fix(model): removed types and introduced new interfaces for MCQ. Refactor name Action->Quizzes
