+++ edweiss-firebase/functions/src/functions/action/creation/createQuiz.ts
@@ -1,15 +1,15 @@
-import Action from 'model/action';
+import Quizzes from 'model/quizzes';
 import { onAuthentifiedCall } from 'utils/firebase';
 import { CollectionOf } from 'utils/firestore';
 import { fail, ok } from 'utils/status';
 
-export const createQuiz = onAuthentifiedCall(Action.Functions.creation.createQuiz, async (userId, args) => {
+export const createQuiz = onAuthentifiedCall(Quizzes.Functions.creation.createQuiz, async (userId, args) => {
     if (args.quiz.exercises.length == 0)
         return fail("empty_quiz");
     if (args.quiz.name.length == 0)
         return fail("invalid_name");
 
-    const quizCollection = CollectionOf<Action.Quiz>(userId + "/quizzes");
+    const quizCollection = CollectionOf<Quizzes.Quiz>(userId + "/quizzes");
 
     const res = await quizCollection.add(args.quiz);
